Move PruneFlag to p2panda-core
