+++ p2panda-engine/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 ciborium = "0.2.2"
 futures-channel = "0.3.30"
 futures-util = { version = "0.3.30", features = ["sink"] }
-p2panda-core = { path = "../p2panda-core" }
+p2panda-core = { path = "../p2panda-core", features = ["prune"] }
 p2panda-store = { path = "../p2panda-store" }
 pin-project = "1.1.5"
 pin-utils = "0.1.0"
