+++ p2panda-engine/src/lib.rs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-pub mod extensions;
 mod macros;
 pub mod operation;
 mod stream;
