+++ p2panda-engine/src/test_utils.rs
@@ -2,10 +2,9 @@
 
 use async_stream::stream;
 use futures_util::Stream;
+use p2panda_core::prune::PruneFlag;
 use p2panda_core::{Body, Extension, Header, PrivateKey, PublicKey, RawOperation};
 use serde::{Deserialize, Serialize};
-
-use crate::extensions::PruneFlag;
 
 #[derive(Clone, Debug, Default, Hash, Eq, PartialEq, Serialize, Deserialize)]
 pub struct StreamName(PublicKey, Option<String>);
